requirements.txt
aws-cdk-lib==2.122.0
constructs>=10.0.0,<11.0.0
cdk-eks-karpenter==1.0.4
aws-cdk.lambda-layer-kubectl-v28==2.2.0